tcc_sigma_agent/requirements.txt
langchain
langchain-google-genai 	#porque vou usar o gemini
langgraph
chromadb
pysigma         #a biblioteca oficial do sigma cli
pyyaml
python-dotenv
